requirements.txt
absl-py==2.1.0
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
asttokens==2.4.1
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cryptography==42.0.2
cssselect2==0.7.0
cycler==0.12.1
decorator==5.1.1
Django==4.2.10
django-cors-headers==4.3.1
djangorestframework==3.14.0
executing==2.0.1
filelock==3.13.1
fonttools==4.48.1
fsspec==2024.2.0
google-auth==2.27.0
google-auth-oauthlib==1.0.0
grpcio==1.60.1
html5lib==1.1
idna==3.6
imageio==2.33.1
importlib-metadata==7.0.1
importlib-resources==6.1.1
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.3
kiwisolver==1.4.5
lxml==5.1.0
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.7.4
matplotlib-inline==0.1.6
mpmath==1.3.0
mysqlclient==2.2.4
networkx==3.1
numpy==1.24.4
oauthlib==3.2.2
opencv-python==4.9.0.80
oscrypto==1.3.0
packaging==23.2
pandas==2.0.3
parso==0.8.3
pickleshare==0.7.5
pillow==10.2.0
prompt-toolkit==3.0.43
protobuf==4.25.2
psutil==5.9.8
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.17.2
pyHanko==0.21.0
pyhanko-certvalidator==0.26.3
PyMuPDF==1.23.21
PyMuPDFb==1.23.9
PyMySQL==1.1.0
pyparsing==3.1.1
pypdf==4.0.1
PyPDF2==3.0.1
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.0.9
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
seaborn==0.13.2
sentry-sdk==1.40.2
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
svglib==1.5.1
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
tinycss2==1.2.1
torch==2.2.0
torchvision==0.17.0
tqdm==4.66.1
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
ultralytics==8.0.20
uritools==4.0.2
urllib3==2.2.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.1
xhtml2pdf==0.2.15
zipp==3.17.0
